+++ mobile/library/kotlin/io/envoyproxy/envoymobile/mocks/MockStream.kt
@@ -32,6 +32,7 @@ class MockStream internal constructor(underlyingStream: MockEnvoyHTTPStream) : S
     override fun getSocketReused(): Boolean { return false }
     override fun getSentByteCount(): Long { return 0 }
     override fun getReceivedByteCount(): Long { return 0 }
+    override fun getResponseFlags(): Long { return 0 }
   }
   /**
    * Closure that will be called when request headers are sent.